+++ packages/support/src/objects/isset.ts
@@ -14,6 +14,10 @@ import type {PropertyPath} from "lodash";
  */
 export function isset<T>(object: T, ...paths: (PropertyPath)[]): boolean
 {
+    if (object === undefined || paths === undefined || paths.length === 0) {
+        return false;
+    }
+
     for (const path of paths) {
         const value = get(object, path);
         if (value === undefined || value === null) {